+++ src/app/(route)/layout.tsx
@@ -1,9 +1,9 @@
 import type { Metadata } from "next";
-import "./globals.css";
-import Providers from "./components/Providers";
+import './globals.css'
+import Providers from "../components/Providers";
 import { getServerSession } from "next-auth";
-import { authOptions } from "./api/auth/[...nextauth]/route";
-import { orbit } from "./utils/fonts";
+import { authOptions } from "../lib/authOptions";
+import { Inter } from 'next/font/google';
 
 export const metadata: Metadata = {
   title: "TO-DO",
@@ -12,6 +12,9 @@ export const metadata: Metadata = {
     icon: "/favicon.png",
   },
 };
+
+const inter = Inter({ subsets: ['latin'] });
+
 
 export default async function RootLayout({
   children,
@@ -22,7 +25,7 @@ export default async function RootLayout({
 
   return (
     <html lang="en">
-      <body className={orbit.className}>
+      <body className={inter.className}>
         <Providers session={session}>
             {children}
         </Providers>
