+++ src/app/(route)/page.tsx
@@ -1,10 +1,15 @@
 "use client";
 import { useRouter } from "next/navigation";
 import { useSession } from "next-auth/react";
+import { useEffect } from "react";
 
 export default function Home() {
   const router = useRouter();
   const { data: session } = useSession();
-  if (session) return router.push("/todo");
-  return router.push("/auth/signin");
+  useEffect(() => {
+    if (session) {
+      return router.push("/todo");
+    }
+    return router.push("/auth/signin");
+  }, []);
 }
