+++ src/app/(route)/todo/page.tsx
@@ -1,10 +1,10 @@
 "use client";
 
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import TodoList from "../components/TodoComponent/Todolist";
-import NewTodo from "../components/TodoComponent/NewTodo";
+import TodoList from "../../components/TodoComponent/Todolist";
+import NewTodo from "../../components/TodoComponent/NewTodo";
 import styled from "@emotion/styled";
-import TodoHeader from "../components/TodoComponent/TodoHeader";
+import TodoHeader from "../../components/TodoComponent/TodoHeader";
 
 const Todo = () => {
   const queryClient = new QueryClient();
