+++ src/app/components/SignOutButton.tsx
@@ -4,12 +4,11 @@ import React from "react";
 import styled from "@emotion/styled";
 import palette from "@/styles/palette";
 import { redirect } from 'next/navigation';
-import { orbit } from "../utils/fonts";
 function SignOutButton() {
   const { data: session } = useSession();
   if (session && session.user) {
     return (
-      <Button className={orbit.className} onClick={() => signOut({ callbackUrl: "/auth/signin" })}>
+      <Button onClick={() => signOut({ callbackUrl: "/auth/signin" })}>
         logout
       </Button>
     );
