--- src/app/utils/fonts.ts
@@ -1,4 +1,0 @@
-import { Orbit, Inter } from 'next/font/google';
-
-export const orbit = Orbit({ preload:false, weight:['400'], });
-export const inter = Inter({ subsets: ['latin'] });
